chore(docs): Optimized array indexing in the log monitor widget's rendering logic
# Project State

## Current Focus
Optimized array indexing in the log monitor widget's rendering logic

## Context
This change follows a pattern of optimizing array indexing operations across the codebase, likely to improve performance in rendering operations.

## Completed
- [x] Removed redundant `as usize` cast in array index calculation
- [x] Simplified index calculation expression

## In Progress
- [x] This is a completed optimization

## Blockers
- None identified

## Next Steps
1. Review other rendering code for similar optimizations
2. Verify performance impact of this change
---
category: chore
scope: docs
---

+++ examples/_cookbook/log_monitor.rs
@@ -211,7 +211,7 @@ impl Widget for LogMonitor {
             let check = if *active { "✓" } else { " " };
             let label = format!("[{}] {}", check, lvl);
             for (j, c) in label.chars().enumerate() {
-                let idx = (2 * w + fx as usize + j) as usize;
+                let idx = 2 * w + fx as usize + j;
                 if idx < p.cells.len() {
                     let fg = if *active { *color } else { t.fg_muted };
                     p.cells[idx] = Cell { char: c, fg, bg: t.surface, style: Styles::empty(), transparent: false, skip: false };
